requirements_training.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements_training.txt requirements_training.in
#
--no-binary pygeos
--no-binary shapely

adal==1.2.7
    # via
    #   azureml-core
    #   msrestazure
affine==2.3.0
    # via
    #   -r requirements_training.in
    #   rasterio
attrs==21.2.0
    # via
    #   fiona
    #   rasterio
azure-common==1.1.27
    # via
    #   azure-graphrbac
    #   azure-mgmt-authorization
    #   azure-mgmt-containerregistry
    #   azure-mgmt-keyvault
    #   azure-mgmt-resource
    #   azure-mgmt-storage
    #   azureml-core
azure-core==1.16.0
    # via azure-mgmt-core
azure-graphrbac==0.61.1
    # via azureml-core
azure-mgmt-authorization==0.61.0
    # via azureml-core
azure-mgmt-containerregistry==8.0.0
    # via azureml-core
azure-mgmt-core==1.3.0
    # via
    #   azure-mgmt-containerregistry
    #   azure-mgmt-keyvault
azure-mgmt-keyvault==9.0.0
    # via azureml-core
azure-mgmt-resource==13.0.0
    # via azureml-core
azure-mgmt-storage==11.2.0
    # via azureml-core
azureml-core==1.32.0
    # via -r requirements_training.in
backports.tempfile==1.0
    # via azureml-core
backports.weakref==1.0.post1
    # via backports.tempfile
certifi==2021.5.30
    # via
    #   fiona
    #   msrest
    #   pyproj
    #   rasterio
    #   requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click==8.0.1
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   rasterio
click-plugins==1.1.1
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
contextlib2==0.6.0.post1
    # via azureml-core
cryptography==3.4.7
    # via
    #   adal
    #   azureml-core
    #   pyopenssl
    #   secretstorage
cycler==0.10.0
    # via matplotlib
docker==4.4.4
    # via azureml-core
fiona==1.8.20
    # via geopandas
geopandas==0.9.0
    # via -r requirements_training.in
idna==2.10
    # via requests
isodate==0.6.0
    # via msrest
jeepney==0.6.0
    # via secretstorage
jmespath==0.10.0
    # via azureml-core
jsonpickle==2.0.0
    # via azureml-core
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.4.2
    # via -r requirements_training.in
msrest==0.6.21
    # via
    #   azure-graphrbac
    #   azure-mgmt-authorization
    #   azure-mgmt-containerregistry
    #   azure-mgmt-keyvault
    #   azure-mgmt-resource
    #   azure-mgmt-storage
    #   azureml-core
    #   msrestazure
msrestazure==0.6.4
    # via
    #   azure-graphrbac
    #   azure-mgmt-authorization
    #   azure-mgmt-resource
    #   azure-mgmt-storage
    #   azureml-core
munch==2.5.0
    # via fiona
ndg-httpsclient==0.5.1
    # via azureml-core
numpy==1.21.0
    # via
    #   -r requirements_training.in
    #   matplotlib
    #   pandas
    #   pygeos
    #   rasterio
    #   snuggs
oauthlib==3.1.1
    # via requests-oauthlib
pandas==1.3.0
    # via geopandas
pathspec==0.8.1
    # via azureml-core
pillow==8.3.0
    # via
    #   -r requirements_training.in
    #   matplotlib
pyasn1==0.4.8
    # via ndg-httpsclient
pycparser==2.20
    # via cffi
pygeos==0.10
    # via -r requirements_training.in
pyjwt==2.1.0
    # via
    #   adal
    #   azureml-core
pyopenssl==20.0.1
    # via
    #   azureml-core
    #   ndg-httpsclient
pyparsing==2.4.7
    # via
    #   matplotlib
    #   snuggs
pyproj==3.1.0
    # via geopandas
python-dateutil==2.8.1
    # via
    #   adal
    #   azureml-core
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   azureml-core
    #   pandas
pyyaml==5.4.1
    # via -r requirements_training.in
rasterio==1.2.6
    # via -r requirements_training.in
requests==2.25.1
    # via
    #   adal
    #   azure-core
    #   azureml-core
    #   docker
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via msrest
ruamel.yaml==0.17.4
    # via azureml-core
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
secretstorage==3.3.1
    # via azureml-core
shapely==1.7.1
    # via
    #   -r requirements_training.in
    #   geopandas
six==1.16.0
    # via
    #   azure-core
    #   cycler
    #   docker
    #   fiona
    #   isodate
    #   msrestazure
    #   munch
    #   pyopenssl
    #   python-dateutil
snuggs==1.4.7
    # via rasterio
torch==1.9.0
    # via -r requirements_training.in
typing-extensions==3.10.0.0
    # via torch
urllib3==1.26.5
    # via
    #   azureml-core
    #   requests
websocket-client==1.1.0
    # via docker

# The following packages are considered to be unsafe in a requirements file:
# setuptools
